refactor: rename to file matcher
