+++ packages/core/src/index.ts
@@ -3,8 +3,8 @@ export { Breakpoints } from './breakpoints'
 export { assignCompositions } from './compositions'
 export { Conditions } from './conditions'
 export { Context } from './context'
-export { FileImport } from './file-import'
-export type { ImportResult } from './file-import'
+export { FileMatcher } from './file-matcher'
+export type { ImportResult } from './file-matcher'
 export { sharedHooks } from './hooks'
 export { ImportMap } from './import-map'
 export { JsxEngine } from './jsx'
@@ -18,6 +18,6 @@ export { StyleDecoder } from './style-decoder'
 export { StyleEncoder } from './style-encoder'
 export { Stylesheet } from './stylesheet'
 export { toCss } from './to-css'
-export type { CssOptions, RecipeContext, RecipeNode, StylesheetContext, TransformResult, ParserOptions } from './types'
+export type { CssOptions, ParserOptions, RecipeContext, RecipeNode, StylesheetContext, TransformResult } from './types'
 export { Utility } from './utility'
 export { messages }
